ai-services/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
scikit-learn==1.4.2
pandas==2.1.4
numpy==1.26.4
scipy==1.11.4
redis==5.0.1
pymongo==4.6.0
python-multipart==0.0.6
httpx==0.25.2
python-dotenv==1.0.0